+++ test/test_runtime_contract.rb
@@ -44,6 +44,7 @@ plasmoid = project_root.join("kde/plasmoids/org.xnix.compatibilitycenter/metadat
   GetRuntimeLiveOwnerGate
   GetRuntimeOwnerSmokePlan
   GetRuntimeMethodParityManifest
+  GetRuntimeWriteGate
   GetCompatibilitySettings
   GetCompatibilitySettingsChangePlan
   GetCompatibilityActionQueue